Move ImageLicensesUpdateForm to Views/ per the rule
Per the views/components rule (#4324): single-use components belong
in app/views/controllers/<controller>/. ImageLicensesUpdateForm only
renders for `images/licenses/edit`, so it moves to:

  Components::ImageLicensesUpdateForm
    -> Views::Controllers::Images::Licenses::Form
  app/components/image_licenses_update_form.rb
    -> app/views/controllers/images/licenses/form.rb
  test/components/image_licenses_update_form_test.rb
    -> test/views/controllers/images/licenses/form_test.rb

Mechanical refactor on top of the existing #4323 work. No behavior
changes; same Superform / FormObject wiring, same rendered HTML.
