+++ app/views/controllers/images/licenses/edit.html.erb
@@ -3,6 +3,6 @@ add_page_title(:image_updater_title.t(user: @user.login))
 container_class(:wide)
 %>
 
-<%= render(Components::ImageLicensesUpdateForm.new(
+<%= render(Views::Controllers::Images::Licenses::Form.new(
              @form,
              action: images_license_updater_path)) %>
